Move Role from membership to roles module
