+++ runtime-modules/membership/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = 'substrate-membership-module'
-version = '1.0.1'
+version = '1.1.0'
 authors = ['Joystream contributors']
 edition = '2018'
