+++ runtime-modules/roles/src/lib.rs
@@ -2,7 +2,10 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod actors;
+mod role_types;
 pub mod traits;
 
 mod mock;
 mod tests;
+
+pub use role_types::Role;
